Double constructor tests were failing due to an incorrect targetType of NodeInfo.Float. Corrected to NodeInfo.Double.
Moves working tests from Debug to New.

DFDL-1119

+++ daffodil-core/src/main/scala/edu/illinois/ncsa/daffodil/dpath/Expression.scala
@@ -1371,7 +1371,7 @@ case class FunctionCallExpression(functionQNameString: String, expressions: List
         XSConverterExpr(functionQNameString, functionQName, args, NodeInfo.Float)
 
       case (RefQName(_, "double", XSD), args) =>
-        XSConverterExpr(functionQNameString, functionQName, args, NodeInfo.Float)
+        XSConverterExpr(functionQNameString, functionQName, args, NodeInfo.Double)
 
       case (RefQName(_, "long", XSD), args) =>
         XSConverterExpr(functionQNameString, functionQName, args, NodeInfo.Long)
